requirements.txt
Django~=3.2.10
Pillow~=8.2.0
django-admin-sortable2==1.0
django-tinymce==3.3.0
